refactor: update/move VideoJS component
