+++ web/components/video/VideoJS/VideoJS.scss
@@ -38,7 +38,7 @@
 }
 
 .vjs-airplay .vjs-icon-placeholder::before {
-  content: url('./airplay.png');
+  content: url('../airplay.png');
 }
 
 .vjs-quality-selector .vjs-icon-placeholder {
